Isolate integration tests
