+++ core/src/test/scala/akka/persistence/jdbc/journal/JdbcJournalPerfSpec.scala
@@ -104,59 +104,6 @@ abstract class JdbcJournalPerfSpec(config: Config, schemaType: SchemaType)
   }
 }
 
-class PostgresJournalPerfSpec extends JdbcJournalPerfSpec(ConfigFactory.load("postgres-application.conf"), Postgres()) {
-  override def eventsCount: Int = 100
-}
-
-class PostgresJournalPerfSpecSharedDb
-    extends JdbcJournalPerfSpec(ConfigFactory.load("postgres-shared-db-application.conf"), Postgres()) {
-  override def eventsCount: Int = 100
-}
-
-class PostgresJournalPerfSpecPhysicalDelete extends PostgresJournalPerfSpec {
-  this.cfg.withValue("jdbc-journal.logicalDelete", ConfigValueFactory.fromAnyRef(false))
-}
-
-class MySQLJournalPerfSpec extends JdbcJournalPerfSpec(ConfigFactory.load("mysql-application.conf"), MySQL()) {
-  override def eventsCount: Int = 100
-}
-
-class MySQLJournalPerfSpecSharedDb
-    extends JdbcJournalPerfSpec(ConfigFactory.load("mysql-shared-db-application.conf"), MySQL()) {
-  override def eventsCount: Int = 100
-}
-
-class MySQLJournalPerfSpecPhysicalDelete extends MySQLJournalPerfSpec {
-  this.cfg.withValue("jdbc-journal.logicalDelete", ConfigValueFactory.fromAnyRef(false))
-}
-
-class OracleJournalPerfSpec extends JdbcJournalPerfSpec(ConfigFactory.load("oracle-application.conf"), Oracle()) {
-  override def eventsCount: Int = 100
-}
-
-class OracleJournalPerfSpecSharedDb
-    extends JdbcJournalPerfSpec(ConfigFactory.load("oracle-shared-db-application.conf"), Oracle()) {
-  override def eventsCount: Int = 100
-}
-
-class OracleJournalPerfSpecPhysicalDelete extends OracleJournalPerfSpec {
-  this.cfg.withValue("jdbc-journal.logicalDelete", ConfigValueFactory.fromAnyRef(false))
-}
-
-class SqlServerJournalPerfSpec
-    extends JdbcJournalPerfSpec(ConfigFactory.load("sqlserver-application.conf"), SqlServer()) {
-  override def eventsCount: Int = 100
-}
-
-class SqlServerJournalPerfSpecSharedDb
-    extends JdbcJournalPerfSpec(ConfigFactory.load("sqlserver-shared-db-application.conf"), SqlServer()) {
-  override def eventsCount: Int = 100
-}
-
-class SqlServerJournalPerfSpecPhysicalDelete extends SqlServerJournalPerfSpec {
-  this.cfg.withValue("jdbc-journal.logicalDelete", ConfigValueFactory.fromAnyRef(false))
-}
-
 class H2JournalPerfSpec extends JdbcJournalPerfSpec(ConfigFactory.load("h2-application.conf"), H2())
 
 class H2JournalPerfSpecSharedDb extends JdbcJournalPerfSpec(ConfigFactory.load("h2-shared-db-application.conf"), H2())
