+++ core/src/test/scala/akka/persistence/jdbc/query/AllPersistenceIdsTest.scala
@@ -53,14 +53,4 @@ abstract class AllPersistenceIdsTest(config: String) extends QueryTestSpec(confi
   }
 }
 
-class PostgresScalaAllPersistenceIdsTest extends AllPersistenceIdsTest("postgres-application.conf") with PostgresCleaner
-
-class MySQLScalaAllPersistenceIdsTest extends AllPersistenceIdsTest("mysql-application.conf") with MysqlCleaner
-
-class OracleScalaAllPersistenceIdsTest extends AllPersistenceIdsTest("oracle-application.conf") with OracleCleaner
-
-class SqlServerScalaAllPersistenceIdsTest
-    extends AllPersistenceIdsTest("sqlserver-application.conf")
-    with SqlServerCleaner
-
 class H2ScalaAllPersistenceIdsTest extends AllPersistenceIdsTest("h2-application.conf") with H2Cleaner
