+++ core/src/test/scala/akka/persistence/jdbc/query/CurrentEventsByPersistenceIdTest.scala
@@ -211,22 +211,6 @@ abstract class CurrentEventsByPersistenceIdTest(config: String) extends QueryTes
 
 // Note: these tests use the shared-db configs, the test for all (so not only current) events use the regular db config
 
-class PostgresScalaCurrentEventsByPersistenceIdTest
-    extends CurrentEventsByPersistenceIdTest("postgres-shared-db-application.conf")
-    with PostgresCleaner
-
-class MySQLScalaCurrentEventsByPersistenceIdTest
-    extends CurrentEventsByPersistenceIdTest("mysql-shared-db-application.conf")
-    with MysqlCleaner
-
-class OracleScalaCurrentEventsByPersistenceIdTest
-    extends CurrentEventsByPersistenceIdTest("oracle-shared-db-application.conf")
-    with OracleCleaner
-
-class SqlServerScalaCurrentEventsByPersistenceIdTest
-    extends CurrentEventsByPersistenceIdTest("sqlserver-shared-db-application.conf")
-    with SqlServerCleaner
-
 class H2ScalaCurrentEventsByPersistenceIdTest
     extends CurrentEventsByPersistenceIdTest("h2-shared-db-application.conf")
     with H2Cleaner
