+++ core/src/test/scala/akka/persistence/jdbc/query/CurrentPersistenceIdsTest.scala
@@ -34,20 +34,4 @@ abstract class CurrentPersistenceIdsTest(config: String) extends QueryTestSpec(c
 
 // Note: these tests use the shared-db configs, the test for all persistence ids use the regular db config
 
-class PostgresScalaCurrentPersistenceIdsTest
-    extends CurrentPersistenceIdsTest("postgres-shared-db-application.conf")
-    with PostgresCleaner
-
-class MySQLScalaCurrentPersistenceIdsTest
-    extends CurrentPersistenceIdsTest("mysql-shared-db-application.conf")
-    with MysqlCleaner
-
-class OracleScalaCurrentPersistenceIdsTest
-    extends CurrentPersistenceIdsTest("oracle-shared-db-application.conf")
-    with OracleCleaner
-
-class SqlServerScalaCurrentPersistenceIdsTest
-    extends CurrentPersistenceIdsTest("sqlserver-application.conf")
-    with SqlServerCleaner
-
 class H2ScalaCurrentPersistenceIdsTest extends CurrentPersistenceIdsTest("h2-shared-db-application.conf") with H2Cleaner
