+++ core/src/test/scala/akka/persistence/jdbc/query/JournalDaoStreamMessagesMemoryTest.scala
@@ -139,19 +139,3 @@ abstract class JournalDaoStreamMessagesMemoryTest(configFile: String)
     }
   }
 }
-
-class PostgresJournalDaoStreamMessagesMemoryTest
-    extends JournalDaoStreamMessagesMemoryTest("postgres-application.conf")
-    with PostgresCleaner
-
-class MySQLJournalDaoStreamMessagesMemoryTest
-    extends JournalDaoStreamMessagesMemoryTest("mysql-application.conf")
-    with MysqlCleaner
-
-class OracleJournalDaoStreamMessagesMemoryTest
-    extends JournalDaoStreamMessagesMemoryTest("oracle-application.conf")
-    with OracleCleaner
-
-class SqlServerJournalDaoStreamMessagesMemoryTest
-    extends JournalDaoStreamMessagesMemoryTest("sqlserver-application.conf")
-    with SqlServerCleaner
